requirements/local.txt
Django==1.8
argparse==1.2.1
django-debug-toolbar==1.3.2
djangorestframework==3.2.1
six==1.9.0
sqlparse==0.1.16
wsgiref==0.1.2
